requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
annotated-types=0.7.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
brotli-python=1.0.9=py310h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
cachetools=5.4.0=pypi_0
certifi=2024.7.4=py310h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py310h06a4308_0
comm=0.2.1=py310h06a4308_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
exceptiongroup=1.2.0=py310h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
gettext=0.22.5=h59595ed_2
gettext-tools=0.22.5=h59595ed_2
google-ai-generativelanguage=0.6.6=pypi_0
google-api-core=2.19.1=pypi_0
google-api-python-client=2.137.0=pypi_0
google-auth=2.32.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-generativeai=0.7.2=pypi_0
googleapis-common-protos=1.63.2=pypi_0
grpcio=1.64.1=pypi_0
grpcio-status=1.62.2=pypi_0
gtts=2.5.1=pyhd8ed1ab_0
httplib2=0.22.0=pypi_0
icu=73.1=h6a678d5_0
idna=3.7=py310h06a4308_0
importlib-metadata=7.0.1=py310h06a4308_0
importlib_metadata=7.0.1=hd3eb1b0_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.25.0=py310h06a4308_0
jedi=0.19.1=py310h06a4308_0
joblib=1.4.2=py310h06a4308_0
jupyter_client=8.6.0=py310h06a4308_0
jupyter_core=5.7.2=py310h06a4308_0
ld_impl_linux-64=2.38=h1181459_1
libasprintf=0.22.5=h661eb56_2
libasprintf-devel=0.22.5=h661eb56_2
libffi=3.4.4=h6a678d5_1
libflac=1.4.3=h59595ed_0
libgcc-ng=14.1.0=h77fa898_0
libgettextpo=0.22.5=h59595ed_2
libgettextpo-devel=0.22.5=h59595ed_2
libgomp=14.1.0=h77fa898_0
libogg=1.3.5=h27cfd23_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=14.1.0=hc0a3c3a_0
libuuid=1.41.5=h5eee18b_0
libxml2=2.10.4=hfdd30dd_2
libxslt=1.1.37=h5eee18b_1
lxml=5.2.1=py310hdbbb534_0
matplotlib-inline=0.1.6=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=py310h06a4308_0
nltk=3.8.1=py310h06a4308_0
openssl=3.0.14=h5eee18b_0
packaging=24.1=py310h06a4308_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=24.0=py310h06a4308_0
platformdirs=3.10.0=py310h06a4308_0
portaudio=19.6.0=h7b6447c_4
prompt-toolkit=3.0.43=py310h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
proto-plus=1.24.0=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pyaudio=0.2.11=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pygments=2.15.1=py310h06a4308_1
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.14=h955ad1f_1
python-dateutil=2.9.0post0=py310h06a4308_2
python-docx=1.1.2=pyhd8ed1ab_0
python-dotenv=1.0.1=pypi_0
pyttsx3=2.90=pypi_0
pyzmq=25.1.2=py310h6a678d5_0
readline=8.2=h5eee18b_0
regex=2023.10.3=py310h5eee18b_0
requests=2.32.2=py310h06a4308_0
rsa=4.9=pypi_0
setuptools=69.5.1=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
speechrecognition=3.10.4=pyhd8ed1ab_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
tk=8.6.14=h39e8969_0
tornado=6.4.1=py310h5eee18b_0
tqdm=4.66.4=py310h2f386ee_0
traitlets=5.14.3=py310h06a4308_0
typing-extensions=4.11.0=py310h06a4308_0
typing_extensions=4.11.0=py310h06a4308_0
tzdata=2024a=h04d1e81_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.2=py310h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.43.0=py310h06a4308_0
xz=5.4.6=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=py310h06a4308_0
zlib=1.2.13=h5eee18b_1
